Move middlewares into subpackage
